test/librados: extract functions using libcommon in test.cc into test_common.cc
Fixes: http://tracker.ceph.com/issues/20175

(cherry picked from commit 82a848c2053ea69ebc6d3ec1003e18921e2d08d2)

Conflicts: since master has switched to cmake. the changes in this
  commit on .am files are ported from counterpart of CMakeLists.txt.
